Extract PHPStanTestCaseTrait and add AbstractBenchCase for phpbench
Extract shared container/config infrastructure from PHPStanTestCase into
PHPStanTestCaseTrait so it can be reused by both PHPUnit tests and phpbench
benchmarks without duplication. Add AbstractBenchCase with runAnalyse method
for writing performance benchmarks.
