+++ phpstan-baseline.neon
@@ -754,10 +754,10 @@ parameters:
 			path: src/Testing/LevelsTestCase.php
 
 		-
-			rawMessage: Anonymous function has an unused use $container.
-			identifier: closure.unusedUse
+			rawMessage: Unused PHPStan\Benchmark\AbstractBenchCase::runAnalyse
+			identifier: shipmonk.deadMethod
 			count: 1
-			path: src/Testing/PHPStanTestCase.php
+			path: tests/PHPStan/Benchmark/AbstractBenchCase.php
 
 		-
 			rawMessage: 'Doing instanceof PHPStan\Type\ConstantScalarType is error-prone and deprecated. Use Type::isConstantScalarValue() or Type::getConstantScalarTypes() or Type::getConstantScalarValues() instead.'
