+++ src/Testing/PHPStanTestCase.php
@@ -8,14 +8,9 @@ use PHPStan\Analyser\Error;
 use PHPStan\Analyser\ScopeFactory;
 use PHPStan\Analyser\TypeSpecifier;
 use PHPStan\BetterReflection\Reflector\Reflector;
-use PHPStan\DependencyInjection\Container;
-use PHPStan\DependencyInjection\ContainerFactory;
 use PHPStan\DependencyInjection\Reflection\ClassReflectionExtensionRegistryProvider;
 use PHPStan\DependencyInjection\Type\ExpressionTypeResolverExtensionRegistryProvider;
 use PHPStan\DependencyInjection\Type\OperatorTypeSpecifyingExtensionRegistryProvider;
-use PHPStan\File\FileHelper;
-use PHPStan\Internal\DirectoryCreator;
-use PHPStan\Internal\DirectoryCreatorException;
 use PHPStan\Node\Printer\ExprPrinter;
 use PHPStan\Parser\Parser;
 use PHPStan\Php\ComposerPhpVersionFactory;
@@ -32,76 +27,17 @@ use PHPStan\Type\TypeAliasResolver;
 use PHPStan\Type\UsefulTypeAliasResolver;
 use PHPUnit\Framework\ExpectationFailedException;
 use PHPUnit\Framework\TestCase;
-use function array_merge;
 use function count;
-use function hash;
 use function implode;
 use function rtrim;
 use function sprintf;
-use function sys_get_temp_dir;
 use const DIRECTORY_SEPARATOR;
-use const PHP_VERSION_ID;
 
 /** @api */
 abstract class PHPStanTestCase extends TestCase
 {
 
-	/** @var array<string, Container> */
-	private static array $containers = [];
-
-	/** @api */
-	public static function getContainer(): Container
-	{
-		$additionalConfigFiles = [__DIR__ . '/TestCase.neon'];
-		foreach (static::getAdditionalConfigFiles() as $configFile) {
-			$additionalConfigFiles[] = $configFile;
-		}
-		$cacheKey = hash('sha256', implode("\n", $additionalConfigFiles));
-
-		if (!isset(self::$containers[$cacheKey])) {
-			$tmpDir = sys_get_temp_dir() . '/phpstan-tests';
-			try {
-				DirectoryCreator::ensureDirectoryExists($tmpDir, 0777);
-			} catch (DirectoryCreatorException $e) {
-				self::fail($e->getMessage());
-			}
-
-			$rootDir = __DIR__ . '/../..';
-			$fileHelper = new FileHelper($rootDir);
-			$rootDir = $fileHelper->normalizePath($rootDir, '/');
-			$containerFactory = new ContainerFactory($rootDir);
-			$container = $containerFactory->create($tmpDir, array_merge([
-				$containerFactory->getConfigDirectory() . '/config.level8.neon',
-			], $additionalConfigFiles), []);
-			self::$containers[$cacheKey] = $container;
-
-			foreach ($container->getParameter('bootstrapFiles') as $bootstrapFile) {
-				(static function (string $file) use ($container): void {
-					require_once $file;
-				})($bootstrapFile);
-			}
-
-			if (PHP_VERSION_ID >= 80000) {
-				require_once __DIR__ . '/../../stubs/runtime/Enum/UnitEnum.php';
-				require_once __DIR__ . '/../../stubs/runtime/Enum/BackedEnum.php';
-				require_once __DIR__ . '/../../stubs/runtime/Enum/ReflectionEnum.php';
-				require_once __DIR__ . '/../../stubs/runtime/Enum/ReflectionEnumUnitCase.php';
-				require_once __DIR__ . '/../../stubs/runtime/Enum/ReflectionEnumBackedCase.php';
-			}
-		} else {
-			ContainerFactory::postInitializeContainer(self::$containers[$cacheKey]);
-		}
-
-		return self::$containers[$cacheKey];
-	}
-
-	/**
-	 * @return string[]
-	 */
-	public static function getAdditionalConfigFiles(): array
-	{
-		return [];
-	}
+	use PHPStanTestCaseTrait;
 
 	public static function getParser(): Parser
 	{
@@ -186,11 +122,6 @@ abstract class PHPStanTestCase extends TestCase
 	protected function shouldTreatPhpDocTypesAsCertain(): bool
 	{
 		return true;
-	}
-
-	public static function getFileHelper(): FileHelper
-	{
-		return self::getContainer()->getByType(FileHelper::class);
 	}
 
 	/**
